Adapt Quests to leo's populate method

+++ Assets/Scripts/Game/NarrativeGenerator/Quests/QuestGrammarNonterminals/Symbol.cs
@@ -7,10 +7,10 @@ namespace Game.NarrativeGenerator
 {
     public interface Symbol 
     {
-        public Dictionary<string, Func<float,float>> nextSymbolChances {get; set;}
+        public Dictionary<string, Func<int,int>> nextSymbolChances {get; set;}
         public string symbolType {get; set;}
         public bool canDrawNext {get; set;}
-        public void SetDictionary( Dictionary<string, Func<float,float>> _nextSymbolChances );
+        public void SetDictionary( Dictionary<string, Func<int,int>> _nextSymbolChances );
         public void SetNextSymbol ( MarkovChain chain );
     }
 }
